fix(build): use cmake to build aws-lc-sys
aws-lc-rs has a bug [1] due to which the AWS_LC_SYS_CFLAGS env var is
ignored in the default "ccBuilder".
When we removed "bindgen" feature, we were switched back to the default
ccBuilder, thus the optimization to not compile AVX512 code was getting
ignored and our binary size increased by 600kB.
This patch reverts to the cmake builder while the cc builder is fixed.

[1]: https://github.com/aws/aws-lc-rs/issues/965

Signed-off-by: Riccardo Mancini <mancio@amazon.com>