fix(tests/startup): measure test time from spawn instead of start
The process startup time is included in spawn, not start. Measuring the
time it takes to start is incorrect. However, let's keep the API calls
in the measurement as this is used as an upper bound to validate the
reported startup time is meaningful.

Also, remove the 400ms sleep as start already waits for the vm to be up.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>