fix: saving/restoring async IO engine transport state
When saving the state of a microVM with one or more block devices backed
by the async IO engine, we need to take a few steps extra steps before
serializing the state to the disk, as we need to make sure that there
aren't any pending io_uring requests that have not been handled by the
kernel yet. For these types of devices that need that we call a
prepare_save() hook before serializing the device state.

If there are indeed pending requests, once we handle them we need to let
the guest know, by adding the corresponding VirtIO descriptors to the
used ring. Moreover, since we use notification suppression, this might
or might not require us to send an interrupt to the guest.

Now, when we save the state of a VirtIO device, we save the device
specific state **and** the transport (MMIO or PCI) state along with it.

There were a few issues with how we were doing the serialization:

1. We were saving the transport state before we run the prepare_save()
   hook. The transport state includes information such as the
   `interrupt_status` in MMIO or `MSI-X config` in PCI. prepare_save()
   in the case of async IO might change this state, so us running it
   after saving the transport state essentially looses information.
2. We were saving the devices states after saving the KVM state. This is
   problematic because, if prepare_save() sends an interrupt to the
   guest we don't save that "pending interrupt" bit of information in
   the snapshot.

These two issues, were making microVMs with block devices backed by
async IO freeze in some cases post snapshot resume, since the guest is
stuck in the kernel waiting for some notification for the device
emulation which never arrives.

Currently, this is only a problem with virtio-block with async IO
engine. The only other device using the prepare_save() hook is currently
virtio-net, but this one doesn't modify any VirtIO state, neither sends
interrupts.

Fix this by ensuring the correct ordering of operations during the
snapshot phase.

Signed-off-by: Babis Chalios <bchalios@amazon.es>